cluster/grpc: add clustered gRPC client and example.
Adds infrastructure for 'smart clients' that can route requests based on
a context key either consistently or round robin, along with server side
hooks for validation/invalidation.

An end to end example exists under the example/ directory, which can be
used as a jumping off point for integration for other services.

This setup currently requires the use of explicit forwarding clients,
but server interceptors/gateways can be built using this infrastructure.
To do so would just be writing an interceptor/forwarder that uses a
ClientConnInterface using the balancer config.