nonce: support reading a new Claimed state.
Add preliminary (backwards compatible) support for a Claimed nonce
state. This state will be used in the future to indicate that a specific
server has 'claimed' a nonce for future use.

In this future system, servers will 'claim' a batch of Available nonces
for use in future transactions. This enables servers to cache nonces to
help reduce the overhead on acquiring a nonce per-intent. The defense
mechanism to abandoned claim is that claims will expire. Healthy servers
will refresh their claimed nonces periodically in the background.